(split)bug 50612 added access check before purging revisions.
http://bugzilla.zimbra.com/show_bug.cgi?id=50612

Copied from Perforce
 Change: 250451

+++ src/java/com/zimbra/cs/mailbox/MailItem.java
@@ -1707,6 +1707,8 @@ public abstract class MailItem implements Comparable<MailItem> {
     }
     
     void purgeRevision(int version, boolean includeOlderRevisions) throws ServiceException {
+        if (!canAccess(ACL.RIGHT_WRITE))
+            throw ServiceException.PERM_DENIED("you do not have the necessary permissions on the item");
         DbMailItem.purgeRevisions(this, version, includeOlderRevisions);
         mRevisions = null;
     }
